Move OperatorHashCollisionsStats to top level
